+++ packages/components/src/components/AccentBox/AccentBox.module.d.scss.ts
@@ -1,0 +1,13 @@
+declare const classNames: {
+  readonly accentBox: "accentBox";
+  readonly gradient: "gradient";
+  readonly neutral: "neutral";
+  readonly blue: "blue";
+  readonly teal: "teal";
+  readonly lilac: "lilac";
+  readonly violet: "violet";
+  readonly green: "green";
+  readonly navy: "navy";
+  readonly icon: "icon";
+};
+export default classNames;