+++ packages/components/src/components/AlertText/AlertText.module.d.scss.ts
@@ -1,0 +1,10 @@
+declare const classNames: {
+  readonly alertText: "alertText";
+  readonly icon: "icon";
+  readonly danger: "danger";
+  readonly success: "success";
+  readonly warning: "warning";
+  readonly info: "info";
+  readonly unavailable: "unavailable";
+};
+export default classNames;