+++ packages/components/src/components/CartesianChart/CartesianChart.module.d.scss.ts
@@ -1,0 +1,9 @@
+declare const classNames: {
+  readonly cartesianChart: "cartesianChart";
+  readonly "flow--illustrated-message--illustrated-message-container": "flow--illustrated-message--illustrated-message-container";
+  readonly emptyViewContainer: "emptyViewContainer";
+  readonly emptyView: "emptyView";
+  readonly "recharts-cartesian-grid": "recharts-cartesian-grid";
+  readonly "recharts-tooltip-wrapper": "recharts-tooltip-wrapper";
+};
+export default classNames;