+++ packages/components/src/components/CartesianChart/components/ChartTooltip/ChartTooltip.module.d.scss.ts
@@ -1,0 +1,7 @@
+declare const classNames: {
+  readonly "recharts-tooltip-wrapper-right": "recharts-tooltip-wrapper-right";
+  readonly "recharts-tooltip-wrapper-left": "recharts-tooltip-wrapper-left";
+  readonly chartTooltip: "chartTooltip";
+  readonly progressBar: "progressBar";
+};
+export default classNames;