+++ packages/components/src/components/DatePicker/components/DateInput/DateInput.module.d.scss.ts
@@ -1,0 +1,8 @@
+declare const classNames: {
+  readonly dateInput: "dateInput";
+  readonly readonly: "readonly";
+  readonly "react-aria-DateInput": "react-aria-DateInput";
+  readonly "react-aria-DateSegment": "react-aria-DateSegment";
+  readonly "flow--button": "flow--button";
+};
+export default classNames;