+++ packages/components/src/components/Flex/Flex.module.d.scss.ts
@@ -1,0 +1,4 @@
+declare const classNames: {
+  readonly flex: "flex";
+};
+export default classNames;