+++ packages/components/src/components/ImageCropper/ImageCropper.module.d.scss.ts
@@ -1,0 +1,9 @@
+declare const classNames: {
+  readonly imageCropper: "imageCropper";
+  readonly cropperContainer: "cropperContainer";
+  readonly unlayered: "unlayered";
+  readonly reactEasyCrop_CropArea: "reactEasyCrop_CropArea";
+  readonly reactEasyCrop_CropAreaGrid: "reactEasyCrop_CropAreaGrid";
+  readonly reactEasyCrop_Container: "reactEasyCrop_Container";
+};
+export default classNames;