+++ packages/components/src/components/Legend/components/LegendItem/LegendItem.module.d.scss.ts
@@ -1,0 +1,5 @@
+declare const classNames: {
+  readonly legendItem: "legendItem";
+  readonly colorSquare: "colorSquare";
+};
+export default classNames;