+++ packages/components/src/components/List/components/Header/components/SearchField/SearchField.module.d.scss.ts
@@ -1,0 +1,4 @@
+declare const classNames: {
+  readonly searchField: "searchField";
+};
+export default classNames;