+++ packages/components/src/components/LoadingSpinner/LoadingSpinner.module.d.scss.ts
@@ -1,0 +1,8 @@
+declare const classNames: {
+  readonly loadingSpinner: "loadingSpinner";
+  readonly light: "light";
+  readonly dark: "dark";
+  readonly "dark-static": "dark-static";
+  readonly "light-static": "light-static";
+};
+export default classNames;