+++ packages/components/src/components/MessageThread/components/MessageSeparator/MessageSeparator.module.d.scss.ts
@@ -1,0 +1,6 @@
+declare const classNames: {
+  readonly messageSeparator: "messageSeparator";
+  readonly separator: "separator";
+  readonly content: "content";
+};
+export default classNames;