+++ packages/components/src/components/Options/Options.module.d.scss.ts
@@ -1,0 +1,5 @@
+declare const classNames: {
+  readonly popover: "popover";
+  readonly options: "options";
+};
+export default classNames;