+++ packages/components/src/components/RadioGroup/RadioGroup.module.d.scss.ts
@@ -1,0 +1,4 @@
+declare const classNames: {
+  readonly radioGroup: "radioGroup";
+};
+export default classNames;