+++ packages/components/src/components/Tabs/components/TabList/TabList.module.d.scss.ts
@@ -1,0 +1,9 @@
+declare const classNames: {
+  readonly tabList: "tabList";
+  readonly activeIndicator: "activeIndicator";
+  readonly animated: "animated";
+  readonly contextMenuButton: "contextMenuButton";
+  readonly titles: "titles";
+  readonly collapsed: "collapsed";
+};
+export default classNames;